fixing bug and adding NEEDED functionallity
for some reason the mouseUp function was never called, leading to a bug breaking the dragging of objects.

I realized that the EditorScene misses out on some very basic functions for controlling it's content from outside. I hope the ones I implemented work.
Also the EditableObjects now have fuction called whenever the Position Updates so for example the numericUpDowns for the current object can change aswell. This function is NOT called when an object is dragged around becuase that's its own function.