add PropertyControl
There is a new Control now which can draw and manage GUI elements that are optimized to change every frame.
Also the TestingForm was updated to have an example implementation of it where it can move objects based on it's center. New events where added to support this kind of interaction.