implement adding pathpoints and simplify selecting by implementing xray picking also some refactoring and improvements to rendering in general
GL_LEQUAL is now the default depth function cause it makes more sense to use
the XRay picking allows you to drag occluded parts of all selected objects which are drawn in a solid yellow color.
A lot of changes needed to be made for it to work and I really hope I can make it less complicated in the future.
When testing everything in legacy gl I found some minor bugs which I fixed in this commit.